+++ gi2_liau7_khoo3/settings/base.py
@@ -9,16 +9,7 @@ https://docs.djangoproject.com/en/1.11/topics/settings/
 For the full list of settings and their values, see
 https://docs.djangoproject.com/en/1.11/ref/settings/
 """
-from django.core.exceptions import ImproperlyConfigured
 import os
-
-def get_env_variable(var_name):
-    try:
-        return os.environ[var_name]
-    except KeyError:
-        error_msg = "Set the %s environment variable" % var_name
-        raise ImproperlyConfigured(error_msg)
-
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
@@ -28,7 +19,6 @@ BASE_DIR = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__fil
 
 # SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = '5hf5)s^_b=m4=cg=_jhpe30mjks1l-v%n3y9b$w(ll5d$hh-ts' 
-#get_env_variable('SECRET_KEY_gi2_liau7_khoo3')
 
 ALLOWED_HOSTS = ['*']
