+++ gi2_liau7_khoo3/settings/local.py
@@ -1,5 +1,3 @@
 from .base import *
 
-# SECRET_KEY = get_env_variable('SOME_SECRET_KEY')
-
 DEBUG = True
