+++ gi2_liau7_khoo3/settings/production.py
@@ -1,4 +1,13 @@
 from .base import *
+from django.core.exceptions import ImproperlyConfigured
 
-# SECRET_KEY = get_env_variable('SOME_SECRET_KEY')
-# DEBUG = False
+
+def get_env_variable(var_name):
+    try:
+        return os.environ[var_name]
+    except KeyError:
+        error_msg = "Set the %s environment variable" % var_name
+        raise ImproperlyConfigured(error_msg)
+
+
+SECRET_KEY = get_env_variable('SECRET_KEY')
